applications/only-menu/backend/pyproject.toml
[tool.poetry]
name = "only-menus"
version = "0.1.0"
description = "A Django Rest Framework-based platform for crowd-sourced restaurant menus"
authors = ["Gaurav Mishra <gaurav.mishra.cx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^4.2"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9.6"
pymongo = "^4.3"
pillow = "^9.5"
transformers = "^4.28"
celery = "^5.2"
redis = "^4.5"
gunicorn = "^20.1"
numpy = "^2.1.0"
drf-spectacular = "^0.27.2"

[tool.poetry.dev-dependencies]
pytest = "^7.3"
pytest-django = "^4.5"
isort = "^5.12"
flake8 = "^6.0"
black = "^23.7"
pylint = "^2.17"
mypy = "^1.3"
django-debug-toolbar = "^4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
ignore_missing_imports = true